+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### 4.12.0 2017-06-05
+
+* Clients can implement `getUserAgentInfo()` to add additional user agent information
+
+### 4.11.0 2017-06-05
+
+* Implement `Countable` for `AttachedObject` (`metadata` and `additional_owners`)
+
+### 4.10.0 2017-05-25
+
+* Add support for login links
+
 ### 4.9.1 2017-05-10
 
 * Fix docs to include arrays on `$id` parameter for retrieve methods